NPE turn timer continues to run when you close the main window move tt to StatusBarPhaseDisplay
